EpsilonPolicy now has a common interface: EpsilonPolicyInterface.
Before EpsilonPolicy was implemented solely in terms of discrete
states. Now it is an interface that can be used both with
discrete states that with Beliefs.
The old EpsilonPolicy is now simply an instantiation of the templated
class EpsilonPolicyInterface. It was also moved inside the MDP
namespace, as all other MDP policies are.