Improved implementation of POMDP Model.
POMDP now inherits from any MDP Model, in order
to provide all of their methods, and to speed up calls
to the underlying MDP as well. Keep in mind that this
is only an implementation, and as long as a class
implements the methods described in POMDP::is_model then
all algorithms are guaranteed to be applicable to it.