New algorithm PBVI. Projecter class. Also changes in the API.
Now I'm trying to be more consistent in the functions that require
passing a size variable for iterating over containers (this is due to
high cost of calling size() in tight loops). PBVI does not yet have a
test however since I'm not sure how to verify that its results make
sense, and the Tiger problem does not seem to be optimal for its belief
finder.