ValueIteration can now be used to solve for any horizon.
In addition it still allows for infinite horizon using the epsilon
bound. The epsilon bound is now kept the same as the one the user sets,
differently from what the MATLAB MDPToolbox does.

ValueIteration will iterate until either the specified horizon is
reached, or convergence within the provided epsilon is reached. The
boolean returned specifies which of the two conditions triggered the
stop.