Introducing WitnessLP_clp
This is an alternative to the lp_solve solver. However, at the moment
this class is not efficient at all, since it does a lot of reallocations
internally since I have not figured out how to force it to keep memory
allocated inside. At the same time, it might be nice for people to have
at least an alternative implementation for the LP.

The CMake build file is already ready to use it, you just need to
uncomment the new parts that were added with this commit.