Removed support for COIN-OR linear programming lib
This is unfortunate, but I could not make the COIN library work for all
the tests. In particular, with the Witness tests the library would just
not find the linear solutions, which completely defeats the point of
having the library at all.

Since lp_solve is also in general very well received and used by other
researchers in this field, I assume that only supporting it will not
have unfortunate consequences.

In any case the plan is to always try to keep the LP interface separate
in order to possibly swap it for another library in the future, if
needed.