Fixed module path in top CMakeLists.txt
CMAKE_SOURCE_DIR refers to the uppermost CMakeLists.txt in the
hierarchy. If this project were to be included by another one, this
would fail to find our modules as the path wouldn't refer to our project
main directory anymore.

We fix this by using PROJECT_SOURCE_DIR in place of the previous
variable.