Fixed QMDP using ValueFunction instead of QFunction.
QMDP was using the MDP ValueFunction instead of the MDP QFunction to
create the PODMP ValueFunction. This was wrong since for each
alphavector/VEntry, all values different from the ones in the
ValueFunction were being initialized to zero.

This deviated the alphavectors, so that the discriminant for when to
change action in middle-beliefs was wrong.

The implementation of QMDP has also been split into a separate static
function, so that it will be usable to create a ValueFunction with the
output of the newly coming FastInformedBound.

I took the time to also add some additional documentation and fix the
API for default creation of ValueFunctions, since I didn't remember the
reason why all POMDP ValueFunctions have a seemingly unuset "zero"
element.