Fix problem with old `ld` version.
There's a bug with older ld versions
(https://sourceware.org/bugzilla/show_bug.cgi?id=22220) which prevents
compilation of the std::getline we use to parse cassandra's files.

This change forces travis to pick the binutils package from a later
distribution (in this case artful, as it's the first one that has an ld
without the bug).

This must be done after having installed everything else, or the system
explodes.