Moved pruner to Prune.hpp in general utils
This refactoring tries to move most, if not all, utilities that operate
on hyperplanes and points outside of the POMDP domain. This is mostly
because in the future we will probably need them for multi-objective
algorithms, and because in any case they are general enough that they do
not need VEntries in order to work (and they never did).

This has the unfortunate effect of requiring a lot more
transform_iterators around, which kind of slows the compilation a bit.
The only hope is that when the ranges TS comes along in C++20, we can
improve the interface to remove all of them.