Added Python wrapper for native generative models.
Using the MDP.GenerativeModelPython it is now possible to wrap native
generative models to use with algorithms such as QLearning or MCTS. This
can be used, for example, to wrap OpenAI gym environments in a very
simple way.

The only thing that is required is a small Python object that contains 5
methods which will be used by the C++ side --- see the docs for more
info.

I also added some MCTS tests in Python so that this functionality can be
tested.