First introduction of basis functions
This is still a work-in-progress. This commit adds the first version of
basis functions, along with their dot-product operation, and the
backProjection operator (which takes a Factored2DMatrix as well).

Some core functions have been modified, some new introduced, and some
are not yet implemented. It's important to notice that most of the
Factored Core utils are probably going to be redundant with Ranges in
C++20. Having some of them named is useful for clarity, but we're
reaching a point that to improve performance we would have to have too
many variants that it would be impractical to handle.