Changed meaning of Table2D and Table3D.
We now uniformly refer to "matrices" as structures using doubles, and
"tables" to structures holding anything else. This required a bit of
renaming, and unfortunately had to alter a couple of API from policies
and experiences, but hopefully nothing too bad.

This change was done mostly to enable to use Eigen matrices (now tables
;) ) for 'long' storage, which can enable us to improve the way
experiences and RLModels work in general.