Cleaned up CPS test
The test as it is is not very informative unfortunately, and it's mostly
a general sanity check. Adding more iterations to try to make it
converge more (and tighten the check) would take quite some time to
complete (some seconds) and I'm not sure whether I want to add the
penalty atm.