Force pick of top parent set
This change is needed since in models with relatively sparse rewards
(i.e. where rewards are only generated by very few nodes), just
iterating randomly over them would slow the propagation of rewards, as
by the time the good node is picked there might be too many constraints
on the output state-action pair to update it correctly. In this way we
are at least always sure that the top priority goes down.