Fix inefficient maxing of must-take local actions.
Since VE selects the action that maximizes the overall sum over local
groups, we don't want to necessarily put the maximum possible value to
local actions which must be taken. This is because then summing would
overflow this values to infinity. This makes it impossible to reliably
select multiple must-take actions at once, as the value of taking just
one or multiple would be the same.

There's still a problem of loss of precision with what we are doing now
(as the values assigned to must-take actions are still so high that
 information from the other action is likely lost), but fixing this
would require setting a "hard cap" on the possible rewards so that the
values we assign are not so big that we lose this. As we don't really
want to ask this information here, we limit ourself to this solution.