Fixed bug with FilterMap, same as LLR
The existing code relied on the FilterMap returning an ordered list to
match the averages with the rules to build. We broke this assumption by
defaulting FilterMap to FasterTrie, which does not return an ordered
list.

This could have been solved in fewer lines by simply forcing FilterMap
to use Trie, but at this point I decided to rewrite the thing to be more
efficient (now that we have the factored matrices), and use the
RollingAverage to keep counts and averages.

We also adapt to UCVE's new API introduced last commit (although the two
changes basically go hand in hand).