Modified methods to accept same-member groups in Q-Function
We now accept (hopefully everywhere, there might still be some methods
which don't take this into account) same-member groups in a
factored bandit's QFunction.

The reason why we want to accept them is that it allows us to exploit
more structure than we would be able to otherwise. For example, consider
a factored bandit where the reward of a group is the sum of the throw of
two separate coins, with two different probabilities of showing heads.
If we know this in advance (and we get the rewards for the coins
separately) we can learn each coin separately faster, and thus perform
better in the bandit. If we had merged the two, we would end up with a
distribution with two peaks, and it would be harder to determine its
mean.

As an addition, we swapped in MOVE and UCVE the tag/value elements in
the Entries so that they are consistent with the other VE
implementations.