Refactored Experience and SparseExperience
This is the first of a series of commits that try to make the API
between Experiences and RollingAverage (which will be renamed shortly to
Experience as well) as uniform as possible.

Given that per-transition sum rewards are not really useful, I have
removed them. I have instead modified the Experience class so that it
directly reports the average of all received rewards for any particular
state-action pair, and it also automatically computes the M2 statistic
for it, just as RollingAverage does. This is done to allow to implement
a ThompsonModel, as we need the std statistic for the recorded rewards.

The matrices for the dense Experience have also been swapped for Eigen
tables/matrices, so that they can be worked with a bit more efficiently
in my internal code.