Updated IO code for Experiences
Given that the API of Experiences has significantly changed, the IO code
for them had to change as well. The new format should be significantly
more readable as it splits the counts separately from the rewards (and
M2s), and also splits each state-action pair data into separate lines.

We also update the test data file used to verify that IO still works.