Removed unused networks to simplify refactoring
I've removed the classes that were not being used, which were only
adding to the complexity and mainteinance with no real reason. It's a
bit sad because having more stuff is always nice, but at the same time
it needs to be maintained and it was becoming a bit of a burden to
handle all classes together in an uniform way.