Introduced DDNGraph class
We now have a new class, DDNGraph, to store the structure of the DDN
once and for all, which removes the need to copy it everywhere. This
commit only contains the basic changes for its introduction; we'll clean
up in the next few commits. No output should have changed. It also makes
reward and transition tables a bit easier to handle, since (1) we don't
need 3 levels of indirections anymore and (2) all the indeces
computations is now done in the new class instead of being repeated
everywhere.

Ideally the factored matrices should also start to use it but I'm not
sure whether it will be easily possible. CPSQueue also has been left
alone since it needs to store additional data for the intermediate
"steps", but I hope I'll be able to simplify it with time.