CPSQueue: fix bug skipping maxA
For some reason I was convinced that jj should have skipped maxA, since
in this part of the code we already decided that we were not going to
get the overall max.

However, not only this is somewhat dumb, as:

- Even if we cannot fully match the overall max, we might still be able
  to match maxA.
- There might still be a second alternative in maxA which is good, which
  is another reason not to skip it.
- In the later stages, there might be no actual alternative to maxA, if
  all other possible action combinations are already taken.

I only realized this bug existed because running the test in debug the
assertion in the assignMatch function was triggering, which means that
basically we were overwriting previous decisions as the jj loop was
ending on an impossible action.

In addition, since this change modifies the behaviour of CPS, we update
its test result number.