CooperativeExperience: compute M2s and average rewards directly.
With this commit CooperativeExperience does the same work as the MDP
Experience; automatically averaging rewards and computing the M2
statistic (so we can later implement a ThompsonModel).

We also update the tests; the CooperativeExperience tests since the
reward function is now different, and also the
CooperativePrioritizedSweeping test. This is due to the new way we
compute the average reward, which has some minor floating point
differences between the old and new way, which results in the final
exact number turning out different.