Introducing ThompsonModel
We finally introduce ThompsonModel, which is a new way to learn a Model
using RL. Instead of simply computing the transition and reward
functions as maximum likelihood, we now sample them from Dirichlet and
Student-t distributions respectively. This should make it unnecessary
(at least for some model-based approaches) to have a separate
exploration policy, as the uncertainty would be already "baked-in" in
the sampled model.