CooperativeModel: added copy constructor
If an explicit or implicit copy was made of a CooperativeModel before
this commit (but after the new DDNGraph was introduced), everything
would break.

This is simply due to the fact that CooperativeModel contains a
recursive reference to itself, and by naively copying the reference
would remain pointed to the original DDNGraph rather than the new,
copied one.