BeliefGenerator: significant speed up
We now keep a bit more information on each belief to avoid as much
redundant work as possible. In addition we split the logic of extending
the reachable beliefs with the one that adds them to the looked up set,
so that we always add the furthest beliefs.

The new logic may end up creating many more beliefs than the accepted
ones (depending on the problem); in that case one could force an overall
limit on how many beliefs are generated, and just stop there.