CPS: Fix Q Update
The old update unfortunately was suboptimal as it didn't allow for
cross-pollination between the different Q-factors, which resulted in
suboptimal decisions.

This new update is more similar to the one performed by SCQL, and has
much better performance in our benchmarks.

We had to increase the value checked for the difference between the true
and approximate QFunction in the test though. The new value is not
really nice; at the same time the new algorithm really does perform
better on larger environments, so we keep it like this.