extractDominatedIncremental: modified API
I've refined the API of this function a bit so that the range after
pruning is much more "organized", we return 4 subranges which contain
the dominated and non-dominated vectors sorted by whether they were in
the old "good" set or in the new ones.

This doesn't really look to be more expensive than before, and has the
advantage that we can optimize new checks on belief points a lot (since
we then would only need to test for increases against the new good
vectors, rather than every vector again).