Utils: Added simple Adam gradient descent implementation.
This is a simple implementation of Adam. It tries to be as
straightforward as possible, and extremely cheap to call and update
(although not to build given the internal Vector allocation required for
the moment estimates).

It currently doesn't have much use in the library, but hopefully when I
get around to pushing GPs it will make more sense. In the meantime it
cannot hurt to have it around in Utils.