Moved Seeder and CassandraParser out of Impl
Seeder was originally in the main Impl folder because it was supposed to
work behind the scenes. However, since I added the `setRootSeed` method,
it has effectively become part of the library's interface.

CassandraParser should have never been there in the first place, but the
Tools folder wasn't there when I wrote it (I think). So now it's in a
better place.