Factored TypeTraits: added QFunctionRuleRange
These concepts are needed for the new maximization selection mechanism.

Before this classes like VariableElimination were able to construct a
FactorGraph from a not-better-identified template "Iterable", which was
assumed to be an iterable of QFunctionRules.

Now we instead introduce the concept of QFunctionRuleRange which
replaces that "Iterable" and adds some type safety. The ranges accept
a range of any class that has the same interface of the appropriate
Bandit/MDP QFunctionRules.

The reason why we check interfaces rather than just checking the exact
class, is that we want to be able to reuse the code that builds graphs
out of Bandit QFunctionRules for the MDP QFunctionRules (which works
equivalently).

Another solution would have just been to make the MDP QFunctionRules
inherit from the Bandit QFunctionRules. This would have worked fine, but
it would have messed with the initialization code of the MDP
QFunctionRules (since the order of the member would have changed from
state/action/value to action/value/state due to inheritance member
ordering).