Added GraphUtils files.
These files are the core of the new mechanism to allow easy swapping
between the maximizer functors we currently have (VariableElimination,
LocalSearch, MaxPlus, ReusingIterativeLocalSearch). It also allows for
extension to new methods in a relatively painless way.

The main mechanism is templatized, so swapping one method for another
will require recompilation, but as the methods are fairly different this
doesn't feel like a terrible tradeoff.

The idea is to provide two new functors: MakeGraph and UpdateGraph.
These are used, respectively, to generate a new graph tuned for a
specific maximizer, and to update the data contained inside that graph
using some form of QFunction.

This division allows for methods which do not destructively operate on
the graphs (like LocalSearch) to save some work and just update their
values, rather than having to construct a new graph from scratch each
time a maximization is needed. Methods that *do* operate destructively
(like VariableElimination), simply don't do any meaningful work in the
MakeGraph and reset everything at each UpdateGraph call.

Additional information about the mechanism is written at the top of the
two files.