Remove member methods to construct graphs.
At the moment I have not removed them from MOVE and UCVE, because they
apply to different settings and cannot be directly swapped with the
other methods. This is not to say that they will not be added in, just
that it's not urgent and can be done in a separate occasion.

Additionally, we made the operator() calls more uniform by removing
additional parameters (when possible). Instead, the parameters have been
added directly to the classes (for example in RILS). This allows the
parameters to be set across maximizations separately, while keeping the
main interface uniform which makes it simpler to swap.