Update the QGreedyPolicies to use the new graph mechanism.
This is the main change of the library, and the main reason why this
whole branch was started. With this change, the QGreedyPolicies can be
effortlessly be configured to use a different maximizer depending on
what is needed.

This is important because before we were only really able to use
VariableElimination. While this is technically the best choice by
default (since VE performs an exact maximization of a factored
QFunction), in large environment it can quickly become too
computationally expensive to be used.

In these cases, it makes a lot of sense to allow an approximate
maximization method (like MaxPlus or LocalSearch) to be used instead.