Update methods to work correctly with the new QGreedyPolicies.
The main change here depends on the fact that now the factored MDP
QGreedyPolicy expects that the QFunction is fully "set up" upon
construction.

This commit moves the QFunction initialization within the initialization
list of the appropriate constructor, before the internal QGreedyPolicies
are built, which allows everything to work correctly.